Renamed component
CreateNewAnimal to AnimalMutation.
