+++ src/components/AnimalMutation/index.tsx
@@ -1,0 +1,2 @@
+import AnimalMutation from "./AnimalMutation";
+export default AnimalMutation;
