--- src/components/CreateNewAnimal/index.tsx
@@ -1,2 +1,0 @@
-import CreateNewAnimal from './CreateNewAnimal';
-export default CreateNewAnimal;
